% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypca.R
\name{chisq_score}
\alias{chisq_score}
\title{Compute chi squared score for PCs}
\usage{
chisq_score(tidypca, newdata)
}
\arguments{
\item{tidypca}{A \code{tidypca} object}

\item{newdata}{Either original data or PCA-transformed}
}
\value{
A data frame with PCs and overall chi squared score
}
\description{
Compute chi squared score for PCs
}
