% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Computes summary statistics for numeric variables}
\usage{
describe(
  df,
  ...,
  add_std_fns = TRUE,
  is_tidy = FALSE,
  tidy_names = ".names",
  tidy_vals = ".x"
)
}
\arguments{
\item{df}{a data frame for which to provide the summary}

\item{...}{"Name = Function" pairs for summary statistics}

\item{add_std_fns}{an indicator whether to add standard descriptive}

\item{is_tidy}{Indicator if data is in tidy format. If so, describe
variables in \code{tidy_names} column using values in \code{tidy_vals}}

\item{tidy_names}{If in tidy format, columns with names and values
statistics (count, number of missing values, mean, standard deviation,
min, 25\%, 50\%, 75\% quantiles, and max)}

\item{tidy_vals}{See \code{tidy_names}}
}
\value{
A tibble with statistics in rows and variables in columns
}
\description{
Computes summary statistics for numeric variables
}
