% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_cols.R
\name{drop_na_cols}
\alias{drop_na_cols}
\title{Drop columns with too many missing values}
\usage{
drop_na_cols(.data, ..., .na_max = 0.25)
}
\arguments{
\item{.data}{A data frame}

\item{...}{tidy-select columns}

\item{.na_max}{Maximum proportion of missing values allowed}
}
\value{
A \code{drop_na_cols} object
}
\description{
Drop columns with too many missing values
}
