% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cached.R
\name{load_cached}
\alias{load_cached}
\title{Load either original or cached data}
\usage{
load_cached(fname, load_fun, write_fun, read_fun, reload = FALSE)
}
\arguments{
\item{fname}{File name}

\item{load_fun}{Function of no arguments to load the data (thunk)}

\item{write_fun}{Function accepting a data frame and file name to write
cached version of the data}

\item{read_fun}{Function accepting one argument, a file name, to read cached
data}

\item{reload}{Whether load fresh or cached data}
}
\value{
Loaded or read data
}
\description{
Load either original or cached data
}
