% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_ncomp_from_recipe.R
\name{pca_ncomp_from_recipe}
\alias{pca_ncomp_from_recipe}
\title{Returns number of PCA components from a PCA recipe step}
\usage{
pca_ncomp_from_recipe(recipe, id = "pca")
}
\arguments{
\item{recipe}{A trained (\code{\link[recipes:prep]{recipes::prep()}}-ed) recipe}

\item{id}{ID of the PCA step}
}
\value{
Number of PCA components or \code{NULL} if a PCA step is not found
}
\description{
Returns number of PCA components from a PCA recipe step
}
