% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypca.R
\name{predict.tidypca}
\alias{predict.tidypca}
\title{Transform values to or from PCA latent space}
\usage{
\method{predict}{tidypca}(object, newdata, to_latent = TRUE, ...)
}
\arguments{
\item{object}{A `tidypca` object}

\item{newdata}{A data frame-ilke data}

\item{to_latent}{if `TRUE`, performs usual transformation to PCs (latent
space). Otherwise, transforms from latent to original variables.}

\item{...}{Unused}
}
\description{
Transform values to or from PCA latent space
}
