% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ravc_cols}
\alias{ravc_cols}
\title{Color-blind friendly colors.}
\format{
An object of class \code{character} of length 4.
}
\usage{
ravc_cols
}
\description{
Color-blind friendly colors.
}
\keyword{datasets}
