% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize data in the data frame}
\usage{
standardize(.data, ..., .loc_fn = mean, .scale_fn = stats::sd)
}
\arguments{
\item{.data}{A data frame}

\item{...}{tidy-select columns}

\item{.loc_fn}{Location function (e.g., `mean` or `median`)}

\item{.scale_fn}{Scale function (e.g., `sd` or `mad`)}
}
\value{
A standardize object
}
\description{
Standardize data in the data frame
}
