% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_max_allowed_na.R
\name{step_select_max_allowed_na}
\alias{step_select_max_allowed_na}
\title{Select columns with not-too-many missing values}
\usage{
step_select_max_allowed_na(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  na_max = 0.05,
  selected_cols = NULL,
  dropped_cols = NULL,
  skip = FALSE,
  id = recipes::rand_id("select_max_allowed_na")
)
}
\arguments{
\item{recipe}{Recipe to which to add the step}

\item{...}{Columns which need to be checked}

\item{role}{For internal use}

\item{trained}{For internal use}

\item{na_max}{Max portion of missing values in a column}

\item{selected_cols}{After training will contain names of selected columns}

\item{dropped_cols}{After training will contain names of dropped columns}

\item{skip}{Whether to skip the step during \code{\link[recipes:bake]{recipes::bake()}}-ing}

\item{id}{Step id}
}
\value{
Updated recipe with the step added
}
\description{
Select columns with not-too-many missing values
}
