% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypca.R
\name{tidypca}
\alias{tidypca}
\title{PCA transfomration on a data frame}
\usage{
tidypca(.data, ..., .center = TRUE, .scale = FALSE, .max_var = 1)
}
\arguments{
\item{.data}{A data frame to apply PCA to}

\item{...}{tidy-select columns to use}

\item{.center}{An indicator whether to center the data prior to PCA}

\item{.scale}{An indicator whether to scale the data prior to PCA}

\item{.max_var}{A parameter contolling the number of PCs: how much of total
variance should the PCs explain.}
}
\value{
A \code{tidypca} (S3) object with fields \code{prcomp}, \code{cols} and \code{ncomp}.
}
\description{
Apply PCA to a data frame. Convenience wrapper over \code{prcomp}
}
